Correções nos endpoints update de loja e produto
Corrigido o bug que quando você realizava o update de produto e/ou loja o mesmo perdia o vinculo com loja e/ou  empreendedor. Implementado a verificação de que caso o empreendedor tenha loja ele não pode ser excluido.

+++ src/main/java/com/api/projetoFinal/services/EmpreendedorService.java
@@ -52,6 +52,9 @@ public class EmpreendedorService {
 
 	public void delete(Integer id) {
 		Empreendedor obj = findById(id);
+		if(obj.getLoja() != null){
+			throw new DataIntegrityViolationException("Não é possível excluir pois você possui uma loja");
+		}
 		repository.deleteById(id);
 
 	}
